Conform the driver to the C standard.
driver.c now contains the following files.
- attach.c
- detach.c
- probe.c

While at it refactored the detach routine.
